test(plugin-agent-orchestrator): consolidate duplicate completion-evidence and terminal-capabilities suites
Closes #28919

Folds each module's src/__tests__ suite into its __tests__/unit/ twin with
zero assertion loss (29 tests total): the #16523 unverified-file-claims
renderer pins and the #9146 device-gating env matrix now live beside the
assembler and structural contracts they pin. Removes the thinner duplicate
files and a drifted duplicate formatOrchestratorCapabilities describe.

AI provider/model: zai / glm-5.3
Client / agent tooling: Hermes Agent
Contribution skill revision: elizaOS/eliza@2bca2475c434e3b57d1b84753087fc73309094ab:packages/skills/skills/contribute-to-eliza
Attribution status: self-reported
— [hermes-t3rpz]
